Refactor js_utils from page_utils
